Make ksmv2 a long-running check
