+++ pkg/collector/runner/runner.go
@@ -8,7 +8,6 @@ package runner
 import (
 	"expvar"
 	"fmt"
-	"strings"
 
 	"strconv"
 	"sync"
@@ -29,29 +28,17 @@ const (
 	stopCheckTimeout time.Duration = 500 * time.Millisecond
 	// Time to wait for all checks to stop
 	stopAllChecksTimeout time.Duration = 2 * time.Second
-	// How long is the first series of check runs we want to log
-	firstRunSeries uint64 = 5
 )
 
 var (
 	// TestWg is used for testing the number of check workers
 	TestWg      sync.WaitGroup
 	runnerStats *expvar.Map
-	checkStats  *runnerCheckStats
 )
 
 func init() {
 	runnerStats = expvar.NewMap("runner")
 	runnerStats.Set("Checks", expvar.Func(expCheckStats))
-	checkStats = &runnerCheckStats{
-		Stats: make(map[string]map[check.ID]*check.Stats),
-	}
-}
-
-// checkStats holds the stats from the running checks
-type runnerCheckStats struct {
-	Stats map[string]map[check.ID]*check.Stats
-	M     sync.RWMutex
 }
 
 // Runner ...
@@ -314,7 +301,7 @@ func (r *Runner) work() {
 			// otherwise only do so if the check is in the scheduler
 			if r.scheduler == nil || r.scheduler.IsCheckScheduled(check.ID()) {
 				mStats, _ := check.GetMetricStats()
-				addWorkStats(check, time.Since(t0), err, warnings, mStats)
+				AddWorkStats(check, time.Since(t0), err, warnings, mStats)
 			}
 		}
 		r.m.Unlock()
@@ -336,86 +323,6 @@ func (r *Runner) work() {
 	}
 
 	log.Debug("Finished processing checks.")
-}
-
-func shouldLog(id check.ID) (doLog bool, lastLog bool) {
-	checkStats.M.RLock()
-	defer checkStats.M.RUnlock()
-
-	var nameFound, idFound bool
-	var s *check.Stats
-
-	loggingFrequency := uint64(config.Datadog.GetInt64("logging_frequency"))
-	name := strings.Split(string(id), ":")[0]
-
-	stats, nameFound := checkStats.Stats[name]
-	if nameFound {
-		s, idFound = stats[id]
-	}
-	// this is the first time we see the check, log it
-	if !idFound {
-		doLog = true
-		lastLog = false
-		return
-	}
-
-	// we log the first firstRunSeries times, then every loggingFrequency times
-	doLog = s.TotalRuns <= firstRunSeries || s.TotalRuns%loggingFrequency == 0
-	// we print a special message when we change logging frequency
-	lastLog = s.TotalRuns == firstRunSeries
-	return
-}
-
-func addWorkStats(c check.Check, execTime time.Duration, err error, warnings []error, mStats map[string]int64) {
-	var s *check.Stats
-	var found bool
-
-	checkStats.M.Lock()
-	log.Tracef("Add stats for %s", string(c.ID()))
-	stats, found := checkStats.Stats[c.String()]
-	if !found {
-		stats = make(map[check.ID]*check.Stats)
-		checkStats.Stats[c.String()] = stats
-	}
-	s, found = stats[c.ID()]
-	if !found {
-		s = check.NewStats(c)
-		stats[c.ID()] = s
-	}
-	checkStats.M.Unlock()
-
-	s.Add(execTime, err, warnings, mStats)
-}
-
-func expCheckStats() interface{} {
-	checkStats.M.RLock()
-	defer checkStats.M.RUnlock()
-
-	return checkStats.Stats
-}
-
-// GetCheckStats returns the check stats map
-func GetCheckStats() map[string]map[check.ID]*check.Stats {
-	checkStats.M.RLock()
-	defer checkStats.M.RUnlock()
-
-	return checkStats.Stats
-}
-
-// RemoveCheckStats removes a check from the check stats map
-func RemoveCheckStats(checkID check.ID) {
-	checkStats.M.Lock()
-	defer checkStats.M.Unlock()
-	log.Debugf("Remove stats for %s", string(checkID))
-
-	checkName := strings.Split(string(checkID), ":")[0]
-	stats, found := checkStats.Stats[checkName]
-	if found {
-		delete(stats, checkID)
-		if len(stats) == 0 {
-			delete(checkStats.Stats, checkName)
-		}
-	}
 }
 
 func getHostname() string {
